Fixes mismatching feedback key for railcom returns when running on a 64-bit processor.

+++ src/openlcb/TractionCvSpace.cxx
@@ -330,7 +330,7 @@ void TractionCvSpace::send(Buffer<dcc::RailcomHubData> *b, unsigned priority)
     if (errorCode_ != ERROR_PENDING)
         return;
     const dcc::Feedback &f = *b->data();
-    if (f.feedbackKey != reinterpret_cast<size_t>(this) || f.channel == 0xff)
+    if (f.feedbackKey != (uint32_t)(reinterpret_cast<size_t>(this)) || f.channel == 0xff)
     {
         // Skip railcom from other packets; also skip the railcom-based
         // occupancy information packets.
